Added support for current date/time and introduced new DayAndTime class
Added ability for the user to set a function for getting the current date and time to the DateAndTime class and use that function in the static member function getCurrentDateAndTime(). This is useful for writing code to get the current date and time using the DateAndTime and DateAndTimeBytes classes in a generic and uniform way.
Added DaysAndTime class which is useful for storing time spans (differences between 2 DateAndTimes).