Additions and Modifications to support Alarms
Added a new Alarm class to allow using temperature-specific Alarms.
Any number of Alarms for temperature ranges can be created and used.
To use Alarms it is necessary for the client to set a GetCurrentDateAndTimeHanlderFunc in DateAndTime's currentDateAndTimeHandler static function pointer data member via the static setGetCurrentDateAndTimeFunction() function.
This serves as a delegate whereby when DateAndTime::getCurrentDateAndTime() is called, the call is forwarded to this delgate function.
Added a new DayAndTime class to better support elapsed times/time spans, and time-of-day functionality.
Additions to DateAndTime to better accommodate Alarm functionality.